+++ src/main/kotlin/apply/application/ApplicantAuthenticationService.kt
@@ -25,4 +25,9 @@ class ApplicantAuthenticationService(
             ?: throw ApplicantAuthenticationException()
         return jwtTokenProvider.createToken(applicant.email)
     }
+
+    fun authenticateEmail(email: String, authenticateCode: String) {
+        val applicant = applicantRepository.findByEmail(email) ?: throw IllegalArgumentException("지원자가 존재하지 않습니다. email: $email")
+        applicant.authenticateEmail(authenticateCode)
+    }
 }
