+++ src/test/kotlin/apply/application/ApplicantServiceTest.kt
@@ -151,34 +151,4 @@ internal class ApplicantServiceTest {
             assertThrows<ApplicantAuthenticationException> { subject() }
         }
     }
-
-    @DisplayName("이메일 사용자 인증 시")
-    @Nested
-    inner class AuthenticateEmail {
-        val applicant = createApplicant()
-
-        @Test
-        fun `인증 코드가 일치한다면 인증된 사용자로 변경한다`() {
-            every { applicantRepository.findByEmail(applicant.email) } returns applicant
-
-            applicantService.authenticateApplicant(applicant.email, applicant.authenticateCode)
-
-            assertThat(applicant.authenticated).isTrue
-        }
-
-        @Test
-        fun `인증 코드가 일치하지 않는다면 예외가 발생한다`() {
-            val applicant = createApplicant()
-            every { applicantRepository.findByEmail(applicant.email) } returns applicant
-
-            assertThrows<ApplicantAuthenticationException> {
-                applicantService.authenticateApplicant(
-                    applicant.email,
-                    "wrong_code"
-                )
-            }
-
-            assertThat(applicant.authenticated).isFalse
-        }
-    }
 }
